Trekker ut komponenter i egne filer
