+++ src/frontend/Sider/Admin/Oppfølging/OppfølgingAdmin.tsx
@@ -3,25 +3,21 @@ import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-import { BodyLong, Button, Heading, Table, VStack } from '@navikt/ds-react';
+import { Button, Heading, Table, VStack } from '@navikt/ds-react';
 
 import { KontrollerOppfølgning } from './KontrollerOppfølgning';
-import { Oppfølging, oppfølgingUtfallTilTekst, årsakKontrollTilTekst } from './oppfølgingTyper';
+import { OppfølgingExpandableRowBody } from './OppfølgingExpandableRowBody';
+import { OppfølgingKontrollertDetaljer } from './OppfølgingKontrollertDetaljer';
+import { Oppfølging } from './oppfølgingTyper';
 import { useApp } from '../../../context/AppContext';
 import DataViewer from '../../../komponenter/DataViewer';
 import { byggHenterRessurs, Ressurs } from '../../../typer/ressurs';
-import { formaterIsoDato, formaterIsoDatoTid, formaterIsoPeriode } from '../../../utils/dato';
-import { aktivitetTypeTilTekst } from '../../Behandling/Inngangsvilkår/Aktivitet/utilsAktivitet';
-import { målgruppeTypeTilTekst } from '../../Behandling/Inngangsvilkår/typer/vilkårperiode/målgruppe';
+import { formaterIsoDato, formaterIsoDatoTid } from '../../../utils/dato';
 import { StønadstypeTag } from '../../Behandling/Venstremeny/Oppsummering/StønadstypeTag';
 
 const Container = styled.div`
     margin: 1rem;
     width: 60rem;
-`;
-
-const Kommentar = styled(BodyLong)`
-    white-space: pre-wrap;
 `;
 
 const WidthMaxContent = styled.div`
@@ -135,72 +131,5 @@ export const OppfølgingTabell = ({ oppfølgingerInit }: { oppfølgingerInit: Op
                 </Table.Body>
             </Table>
         </Container>
-    );
-};
-
-const OppfølgingExpandableRowBody = ({ oppfølging }: { oppfølging: Oppfølging }) => {
-    return (
-        <Table size={'small'}>
-            <Table.Header>
-                <Table.Row>
-                    <Table.HeaderCell scope={'col'}>Periode</Table.HeaderCell>
-                    <Table.HeaderCell scope={'col'}>Målgruppe</Table.HeaderCell>
-                    <Table.HeaderCell scope={'col'}>Aktivitet</Table.HeaderCell>
-                    <Table.HeaderCell scope={'col'}>Endring målgruppe</Table.HeaderCell>
-                    <Table.HeaderCell scope={'col'}>Endring aktivitet</Table.HeaderCell>
-                </Table.Row>
-            </Table.Header>
-            <Table.Body>
-                {oppfølging.data.perioderTilKontroll.map((periode, index) => (
-                    <Table.Row key={index}>
-                        <Table.DataCell>
-                            {formaterIsoPeriode(periode.fom, periode.tom)}
-                        </Table.DataCell>
-                        <Table.DataCell>{målgruppeTypeTilTekst(periode.målgruppe)}</Table.DataCell>
-                        <Table.DataCell>{aktivitetTypeTilTekst(periode.aktivitet)}</Table.DataCell>
-                        <Table.DataCell>
-                            <VStack>
-                                {periode.endringMålgruppe.map((endring) => (
-                                    <span key={endring.årsak}>
-                                        {årsakKontrollTilTekst[endring.årsak]}{' '}
-                                        {endring.fom && formaterIsoDato(endring.fom)}
-                                        {endring.tom && formaterIsoDato(endring.tom)}
-                                    </span>
-                                ))}
-                            </VStack>
-                        </Table.DataCell>
-                        <Table.DataCell>
-                            <VStack>
-                                {periode.endringAktivitet.map((endring) => (
-                                    <span key={endring.årsak}>
-                                        {årsakKontrollTilTekst[endring.årsak]}{' '}
-                                        {endring.fom && formaterIsoDato(endring.fom)}
-                                        {endring.tom && formaterIsoDato(endring.tom)}
-                                    </span>
-                                ))}
-                            </VStack>
-                        </Table.DataCell>
-                    </Table.Row>
-                ))}
-            </Table.Body>
-        </Table>
-    );
-};
-
-export const OppfølgingKontrollertDetaljer = ({
-    kontrollert,
-}: {
-    kontrollert: NonNullable<Oppfølging['kontrollert']>;
-}) => {
-    return (
-        <VStack>
-            <span>
-                Kontrollert: {formaterIsoDatoTid(kontrollert.tidspunkt)} av{' '}
-                {kontrollert.saksbehandler}
-            </span>
-            <span>Utfall: {oppfølgingUtfallTilTekst[kontrollert.utfall]}</span>
-            {kontrollert.kommentar && <Kommentar>Kommentar: {kontrollert.kommentar}</Kommentar>}
-            {!kontrollert.kommentar && <span>Ingen kommentar</span>}
-        </VStack>
     );
 };
